third_party/blink/web_tests/fast/dom/custom/document-register-namespace-expected.txt

CONSOLE WARNING: line 24: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Using document.registerElement() for extending HTML and non-HTML elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS html1.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS html2.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS html3.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS html4.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS notHTML.namespaceURI is "http://www.example.com/"
PASS notHTML instanceof CustomHTMLElement is false
PASS notHTML instanceof HTMLElement is false
PASS svg1.tagName is "g"
PASS svg1 instanceof SVGGElement is true
PASS svg1 instanceof CustomSVGElement is true
PASS svg1.namespaceURI is "http://www.w3.org/2000/svg"
PASS svg2.tagName is "svg-foo"
PASS svg2.namespaceURI is "http://www.w3.org/2000/svg"
PASS svg3.tagName is "svg-foo"
PASS svg3.namespaceURI is "http://www.w3.org/2000/svg"
PASS notSVG1.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS notSVG1 instanceof CustomSVGElement is false
PASS notSVG1 instanceof HTMLUnknownElement is false
PASS notSVG1 instanceof HTMLElement is true
PASS Object.getPrototypeOf(notSVG1) is HTMLElement.prototype
PASS notSVG2.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS notSVG2 instanceof CustomSVGElement is false
PASS notSVG2 instanceof HTMLUnknownElement is false
PASS notSVG2 instanceof HTMLElement is true
PASS Object.getPrototypeOf(notSVG2) is HTMLElement.prototype
PASS successfullyParsed is true

TEST COMPLETE
